data_prepare/requirements.txt
numpy
pandas
tqdm
requests==2.32.3
python-dateutil==2.8.2